pyproject.toml
[project]
name = "tinyagent"
version = "0.2.0"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Curt Park", email = "www.jwpark.co.kr@gmail.com" },
]

dependencies = [
    "openai >= 2.28.0",
    "python-dotenv >= 1.1.0",
    "pyyaml >= 6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["run.py", "config"]


[project.scripts]
tinyagent = "run:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.14.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 188
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=run",
    "--cov-report=term-missing",
    "--cov-report=html",
]